Remove the branding stuff, focus only on the lockfile for now.
