+++ plugin-gradle/src/main/java/dev/equo/ide/gradle/EquoIdeExtension.java
@@ -13,25 +13,17 @@
  *******************************************************************************/
 package dev.equo.ide.gradle;
 
-import dev.equo.ide.BrandingIdeHook;
 import dev.equo.ide.IdeHook;
 import dev.equo.solstice.p2.P2Client;
 import dev.equo.solstice.p2.P2Model;
 import dev.equo.solstice.p2.P2Query;
-import org.gradle.api.Project;
 
 /** The DSL inside the equoIde block. */
 public class EquoIdeExtension extends P2ModelDsl {
-	private final Project project;
 	public boolean useAtomos = true;
 	private final IdeHook.List ideHooks = new IdeHook.List();
-	private final BrandingIdeHook branding = new BrandingIdeHook();
 
-	public EquoIdeExtension(Project project) {
-		this.project = project;
-		useAtomos = true;
-		ideHooks.add(branding);
-	}
+	public EquoIdeExtension() {}
 
 	private static void setToDefault(P2Model model) {
 		model.addP2Repo("https://download.eclipse.org/eclipse/updates/4.26/");
